Update app.js
Review 2:

This code seems to operate just fine when I tried it.  You took a little different approach than I did and didn't follow the book as much which is impressive.  You organized your code very well so it is efficient.  You were able to circumvent the error I spent a long time dealing with by not following the book so much. The one issue I had still was the error I got when I tried to access the camera in the app.  I am assuming that is because it is not running on an actual mobile device and the emulators don't access the computer webcams.  However, this would be a cool thing to figure out.  I also like that you put comments throughout your code.  I failed to do this and now I feel bad for whoever is reviewing my code. Despite organizing it efficiently, I think the order you laid out your code is somewhat confusing.  This isn't really an issue right now since you are the only one editing this code, but in a job this may create challenges.  I know this is supposed to have a lot of constructive criticism but the only things I can really find to pick at is lack of creativity with the layout and colors.  I struggle with this as well but it may be something you want to consider in later assignments and especially on the final project.  All in all nice job.